+++ raster/src/main/scala/geotrellis/raster/mapalgebra/focal/Implicits.scala
@@ -17,10 +17,26 @@
 package geotrellis.raster.mapalgebra.focal
 
 import geotrellis.raster.{BufferTile, Tile}
+import shapeless.=:!=
 
 object Implicits extends Implicits
 
 trait Implicits {
-  implicit class withTileFocalMethods(val self: Tile) extends FocalMethods
-  implicit class withBufferTileFocalMethods(val self: BufferTile) extends BufferTileFocalMethods
+  /**
+   * TODO: think of a better way of organizing implicits with regards to possible Tile successors
+   *
+   * Apply withTileFocalMethods to all successor of a [[Tile]] type but [[BufferTile]].
+   * It is required to disambiguate withBufferTileFocalMethods implicit.
+   */
+  implicit class withTileFocalMethods[T](tile: T)(implicit t: T <:< Tile, nbt: T =:!= BufferTile) extends FocalMethods {
+    val self: Tile = t(tile)
+  }
+
+  /**
+   * [[BufferTile]] can't have successors (it is a case class).
+   * Limit this implicit to be applied to the [[BufferTile]] type only.
+   */
+  implicit class withBufferTileFocalMethods[T](tile: T)(implicit bt: T =:= BufferTile) extends BufferTileFocalMethods {
+    val self: BufferTile = bt(tile)
+  }
 }